chore(dependencies): Update most.js dep to 1.0 and fix typing
The latest version of most.js is 1.0.0, and the peer dependency was updated to reflect that. The
update broke the tests because most.js removed its definition of the Promise type, and an external
type dependency resolved the issue.

most-proxy now relies on an external Promise definition rather than the one in the previous version
of most.